Add files via upload
Figure 1, shows a plot of earthquake magnitude versus the depth of the focal point of the earthquake. Analytical solutions in solid mechanics (Sadd. et. al.) tell us that the stress for a load in an infinite medium would take the form of (a/r^4), where r is the distance from the source and a is a constant coefficient. The stress would be expected to scale with the magnitude of the earthquake by a base-10 logarithmic scale. The red points represent earthquake data over the last 100+ years and the blue line indicates an analytical function -log(1000/r^4). Note the negative sign since a greater depth into the crust is actually closer to the source. One can see that although the fit is not perfect, there is at least a qualitative agreement with the analytical model.